@@ -3,12 +3,10 @@
  *******************************************************************************/
 package ca.footeware.web.tests;
 
-import org.assertj.core.api.Assertions;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import ca.footeware.web.controllers.ImageController;
 import ca.footeware.web.controllers.JokeController;
@@ -17,7 +15,6 @@ import ca.footeware.web.controllers.JokeController;
  * @author craig
  *
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest
 public class ApplicationTests {
 
@@ -32,7 +29,7 @@ public class ApplicationTests {
 	 */
 	@Test
 	public void imageControllerLoads() {
-		Assertions.assertThat(imageController).isNotNull();
+		Assertions.assertTrue(imageController != null);
 	}
 
 	/**
@@ -40,7 +37,7 @@ public class ApplicationTests {
 	 */
 	@Test
 	public void jokeControllerLoads() {
-		Assertions.assertThat(jokeController).isNotNull();
+		Assertions.assertTrue(jokeController != null);
 	}
 
 }
