@@ -9,14 +9,12 @@ import java.io.IOException;
 
 import javax.imageio.ImageIO;
 
-import org.junit.Assert;
-import org.junit.Test;
-import org.junit.runner.RunWith;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.boot.test.web.client.TestRestTemplate;
-import org.springframework.test.context.junit4.SpringRunner;
 
 import ca.footeware.web.controllers.ImageController;
 
@@ -25,7 +23,6 @@ import ca.footeware.web.controllers.ImageController;
  * 
  * @author Footeware.ca
  */
-@RunWith(SpringRunner.class)
 @SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
 public class ImageControllerITTests {
 
@@ -42,10 +39,11 @@ public class ImageControllerITTests {
 	@Test
 	public void testGetGalleries() {
 		String page = template.getForObject("/gallery/", String.class);
-		Assert.assertTrue("Should have been sent to login page.",
-				page.contains("<form action=\"/login\" method=\"post\">"));
+		Assertions.assertTrue(page.contains("<form action=\"/login\" method=\"post\">"),
+				"Should have been sent to login page.");
 		page = template.withBasicAuth(USERNAME, PASSWORD).getForObject("/gallery/", String.class);
-		Assert.assertTrue("Should have listed 'gallery1'.", page.contains("href=\"/gallery/gallery1\">gallery1</a>"));
+		Assertions.assertTrue(page.contains("href=\"/gallery/gallery1\">gallery1</a>"),
+				"Should have listed 'gallery1'.");
 	}
 
 	/**
@@ -55,17 +53,17 @@ public class ImageControllerITTests {
 	@Test
 	public void testGetGallery() {
 		String page = template.getForObject("/gallery/gallery1/", String.class);
-		Assert.assertTrue("Should have been sent to login page.",
-				page.contains("<form action=\"/login\" method=\"post\">"));
+		Assertions.assertTrue(page.contains("<form action=\"/login\" method=\"post\">"),
+				"Should have been sent to login page.");
 		page = template.withBasicAuth(USERNAME, PASSWORD).getForObject("/gallery/gallery1/", String.class);
-		Assert.assertTrue("Should have been sent to gallery1 page.",
-				page.contains("<li class=\"active\"><a href=\"/gallery\">Gallery</a></li>"));
-		Assert.assertTrue("Horizontal thumbnail not displayed.",
-				page.contains("href=\"/gallery/gallery1/test-image-horizontal.png\""));
-		Assert.assertTrue("Vertical thumbnail not displayed.",
-				page.contains("href=\"/gallery/gallery1/test-image-vertical.png\""));
-		Assert.assertTrue("Square thumbnail not displayed.",
-				page.contains("href=\"/gallery/gallery1/test-image-square.png\""));
+		Assertions.assertTrue(page.contains("<li class=\"active\"><a href=\"/gallery\">Gallery</a></li>"),
+				"Should have been sent to gallery1 page.");
+		Assertions.assertTrue(page.contains("href=\"/gallery/gallery1/test-image-horizontal.png\""),
+				"Horizontal thumbnail not displayed.");
+		Assertions.assertTrue(page.contains("href=\"/gallery/gallery1/test-image-vertical.png\""),
+				"Vertical thumbnail not displayed.");
+		Assertions.assertTrue(page.contains("href=\"/gallery/gallery1/test-image-square.png\""),
+				"Square thumbnail not displayed.");
 	}
 
 	/**
@@ -79,25 +77,25 @@ public class ImageControllerITTests {
 		byte[] bytes = template.withBasicAuth(USERNAME, PASSWORD)
 				.getForObject("/gallery/gallery1/test-image-horizontal.png", byte[].class);
 		BufferedImage image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 1920, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 1241, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 1920, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 1241, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD).getForObject("/gallery/gallery1/test-image-vertical.png",
 				byte[].class);
 		image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 1241, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 1920, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 1241, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 1920, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD).getForObject("/gallery/gallery1/test-image-square.png",
 				byte[].class);
 		image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 1920, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 1920, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 1920, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 1920, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD).getForObject("/gallery/gallery1/test-image-bad.png",
 				byte[].class);
-		Assert.assertTrue("Should have been an error message for image with bad name.",
-				new String(bytes).contains("gallery1/test-image-bad.png not found."));
+		Assertions.assertTrue(new String(bytes).contains("gallery1/test-image-bad.png not found."),
+				"Should have been an error message for image with bad name.");
 	}
 
 	/**
@@ -110,29 +108,29 @@ public class ImageControllerITTests {
 	public void testGetThumbnail() throws IOException {
 		byte[] bytes = template.withBasicAuth(USERNAME, PASSWORD)
 				.getForObject("/gallery/thumbnails/gallery1/test-image-horizontal.png", byte[].class);
-		Assert.assertEquals("Wrong number of bytes for thumbnail.", 2781, bytes.length);
+		Assertions.assertEquals(bytes.length, 2781, "Wrong number of bytes for thumbnail.");
 		BufferedImage image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 150, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 97, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 150, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 97, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD)
 				.getForObject("/gallery/thumbnails/gallery1/test-image-vertical.png", byte[].class);
-		Assert.assertEquals("Wrong number of bytes for thumbnail.", 1395, bytes.length);
+		Assertions.assertEquals(bytes.length, 1395, "Wrong number of bytes for thumbnail.");
 		image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 97, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 150, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 97, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 150, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD)
 				.getForObject("/gallery/thumbnails/gallery1/test-image-square.png", byte[].class);
-		Assert.assertEquals("Wrong number of bytes for thumbnail.", 1617, bytes.length);
+		Assertions.assertEquals(bytes.length, 1617, "Wrong number of bytes for thumbnail.");
 		image = ImageIO.read(new ByteArrayInputStream(bytes));
-		Assert.assertEquals("Image wrong width.", 150, image.getWidth());
-		Assert.assertEquals("Image wrong height.", 150, image.getHeight());
+		Assertions.assertEquals(image.getWidth(), 150, "Image wrong width.");
+		Assertions.assertEquals(image.getHeight(), 150, "Image wrong height.");
 
 		bytes = template.withBasicAuth(USERNAME, PASSWORD)
 				.getForObject("/gallery/thumbnails/gallery1/test-image-bad.png", byte[].class);
-		Assert.assertTrue("Should have been an error message for thumbnail with bad name.",
-				new String(bytes).contains("gallery1/test-image-bad.png not found."));
+		Assertions.assertTrue(new String(bytes).contains("gallery1/test-image-bad.png not found."),
+				"Should have been an error message for thumbnail with bad name.");
 	}
 
 }
